+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Fixed
 * Fixed a potential crash in xc3_viewer when drawing empty skeletons with the `--bones` option.
+* Fixed an issue where wimdo models would not load the `.chr` or `.arc` file with the full base skeleton.
 
 ## 0.16.0 - 2025-01-03
 ### Added